Support custom tsig names inside tsigSecret map
Previously the tsigSecret variable on the Server object was keyed on tsig name.
This means that if you have multiple zones using the same server, it is not possible
for each of the zones to have the same tsig name. By specifying a TsigKeyNameBuilder
function, the server can now differentiate between zones even if they have the same tsig name
without requiring the client to change anything.