Make ExchangeWithConnContext public
If we want to use a custom mechanism of obtaining a Conn that doesn't
match the net.Dialer type, but retain the timeout behaviour from
ExchangeContext, there was previously no way to accomplish this.

This PR makes the underlying ExchangeWithConnContext function public,
which allows this behaviour.

Now that #1454 is merged, there is no longer any interaction between the
context provided and the singleflight behaviour, so I removed the
comment from ExchangeWithConn.

Signed-off-by: Andrew Dunham <andrew@du.nham.ca>