Avoid using strings.Split (#1501)
* Avoid using strings.Split

strings.Split has to allocate for the return slice. This allocation
was wasteful in ever case it was used in this library.

Instead we use the new strings.Cut and other string manipulation where
appropriate. This tends to lead to cleaner and more readable code in
addition to the benefits this has on the garbage collector.

* Further simplify structTag in the msg_generate.go

This doesn't need to call strings.TrimPrefix twice.