Refactor DNS message unpacking behaviour
We keep some of the existing weird behaviour (mostly of not erroring
when we should), but refactor it so it's far more obvious and things
are cleaner.

We leave a number of TODOs to address some of the weirder choices. They
would also allow for some other improvements to the code if we can make
those changes.