Avoid passing *cryptobyte.String through an interface
In some places this results in more readable code, in other places less.
What we gain though is the same number of allocations that we started
with. In other words, we haven't made things worse!

Without this change, every time we pass *cryptobyte.String into an
interface it is forced onto the heap and we get a nasty allocation.

It does require us to check that all the data has been consumed at
the end of each `unpack([]byte) error` call instead of having a
single check. Ultimately this isn't too bad.

We could change unpack to return whatever is left over. We don't
do this for semantic reasons. unpack is required to consume the
entirety of data (as we know it's length ahead of time). Asking
it to return whatever is left for the sole purpose of checking
it's nothing is silly. In some places we pass a pointer or do
return the rest. In these cases, it's because we don't know
ahead of time how long something will be.