Un-genericfy unpackCounted
While it's nice having this as a single function rather than one for
Question and one for RR, using generics and passing the unpack function
causes the *cryptobyte.String to escape. That leaves us with some
rather ugly workarounds. Instead just bite the bullet and write it
twice. If we chose to remove the msg.Empty() check in the future, then
this doesn't end up too bad at all.