Use semantically sensible error messages when unpacking
This fixes a CoreDNS test failure
(github.com/coredns/coredns/plugin/pkg/proxy.TestCoreDNSOverflow) that
was testing behaviour that is sensitive to the specific error message
returned when unpacking overflowing messages or that it is ErrBuf.

It also ensures we correctly distinguish between invalid record data
and invalid message framing. Ideally the `unpack(...) error` methods
(on RR, EDNS0 and SVCBKeyValue) would actually be `unpack(...) bool`
and we'd allow ourselves more freedom on how we report unpacking
failures.